refactor(dao): update and relocate route-related DAOs
- Updated methods in `ActivityRecordDao`, `NearbyServiceDao`, `PointOfInterestDao`, and `RouteDao`.
- Relocated files to `route/main` and `route/reference` packages for better organization.
